backend/requirements.txt
asgiref==3.6.0
Django==3.2.19
psycopg2==2.9.6
pytz==2023.3
typing_extensions==4.5.0
djoser==2.1.0
pillow==9.5.0
sorl-thumbnail==12.9.0
djangorestframework==3.14.0
django-filter==23.2
python-dotenv==0.21.1
requests==2.31.0